refactor routes,contollers,utils
